[FLINK-13071][table-planner-blink] minor update based on review comment.
