+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/functions/utils/UserDefinedFunctionUtils.scala
@@ -609,8 +609,6 @@ object UserDefinedFunctionUtils {
       function: ScalarFunction,
       arguments: Array[AnyRef],
       argTypes: Array[LogicalType]): DataType = {
-//    val userDefinedTypeInfo = function.getResultType(
-//      arguments, getEvalMethodSignature(function, argTypes))
     val userDefinedTypeInfo = function.getResultType(getEvalMethodSignature(function, argTypes))
     if (userDefinedTypeInfo != null) {
       fromLegacyInfoToDataType(userDefinedTypeInfo)
